streamlit/requirements.txt
streamlit==1.29.0
psycopg2-binary==2.9.9
pandas==2.1.4
plotly==5.18.0
httpx==0.25.2
python-dotenv==1.0.0
